Move capability client APIs to dedicated file
