+++ client/client.go
@@ -127,42 +127,3 @@ func (client *Client) SysInfo() (*SysInfo, error) {
 
 	return &sysInfo, nil
 }
-
-// Capability holds information about a capability that a snap may request
-// from a snappy system to do its job while running on it.
-type Capability struct {
-	// Name is a key that identifies the capability. It must be unique within
-	// its context, which may be either a snap or a snappy runtime.
-	Name string `json:"name"`
-	// Label provides an optional title for the capability to help a human tell
-	// which physical device this capability is referring to. It might say
-	// "Front USB", or "Green Serial Port", for example.
-	Label string `json:"label"`
-	// Type defines the type of this capability. The capability type defines
-	// the behavior allowed and expected from providers and consumers of that
-	// capability, and also which information should be exchanged by these
-	// parties.
-	Type string `json:"type"`
-	// Attrs are key-value pairs that provide type-specific capability details.
-	Attrs map[string]string `json:"attrs,omitempty"`
-}
-
-// Capabilities returns the capabilities currently available for snaps to consume.
-func (client *Client) Capabilities() (map[string]Capability, error) {
-	var rsp response
-	if err := client.do("GET", "/1.0/capabilities", nil, &rsp); err != nil {
-		return nil, err
-	}
-	if rsp.Type == "error" {
-		// TODO: handle structured errors
-		return nil, fmt.Errorf("cannot obtain capabilities: %s", rsp.Status)
-	}
-	if rsp.Type != "sync" {
-		return nil, fmt.Errorf("cannot obtain capabilities: expected sync response, got %s", rsp.Type)
-	}
-	var result map[string]map[string]Capability
-	if err := json.Unmarshal(rsp.Result, &result); err != nil {
-		return nil, fmt.Errorf("cannot obtain capabilities: failed to unmarshal response: %v", err)
-	}
-	return result["capabilities"], nil
-}
